# docs/src/apps/swagger.yaml
info:
  description: Nesis API
  version: 0.1.1
  title: Nesis API
paths:
  /v1/apps:
    get:
      parameters:
      - in: header
        name: Authorization
        schema:
          type: string
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Apps'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    post:
      summary: Creates a new app.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppReq'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppPostRes'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
openapi: 3.0.2
components:
  schemas:
    AppReq:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AppRes:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        id:
          type: string
        enabled:
          type: boolean
        create_date:
          type: string
          format: null
          pattern: null
    Apps:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppRes'
        count:
          type: integer
...
    Message:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    AppPostRes:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        id:
          type: string
        enabled:
          type: boolean
        create_date:
          type: string
          format: null
          pattern: null
        secret:
          type: string